+++ include/sir/SIRSchedulerBase.h
@@ -28,9 +28,9 @@ public:
   const unsigned EntrySlot;
   static const unsigned MaxSlot = UINT16_MAX >> 2;
   struct TimeFrame {
-    float ASAP, ALAP;
+    unsigned ASAP, ALAP;
 
-    TimeFrame(float ASAP = UINT32_MAX, float ALAP = 0)
+    TimeFrame(unsigned ASAP = UINT32_MAX, unsigned ALAP = 0)
       : ASAP(ASAP), ALAP(ALAP) {}
 
     TimeFrame(const TimeFrame &RHS) : ASAP(RHS.ASAP), ALAP(RHS.ALAP) {}
@@ -56,7 +56,7 @@ public:
 protected:
   // MII in modulo schedule.
   unsigned MII;
-  float CriticalPathEnd;
+  unsigned CriticalPathEnd;
 
   typedef std::map<const SIRSchedUnit *, TimeFrame> TFMapTy;
   TFMapTy SUnitToTF;
@@ -75,8 +75,8 @@ public:
     return G.getSUsInBB(BB);
   }
 
-  float calculateASAP(const SIRSchedUnit *A) const;
-  float calculateALAP(const SIRSchedUnit *A) const;
+  unsigned calculateASAP(const SIRSchedUnit *A) const;
+  unsigned calculateALAP(const SIRSchedUnit *A) const;
   TimeFrame calculateTimeFrame(const SIRSchedUnit *A) const {
     return TimeFrame(calculateASAP(A), calculateALAP(A));
   }
@@ -97,13 +97,13 @@ public:
   typedef SIRSchedGraph::reverse_iterator reverse_iterator;
   typedef SIRSchedGraph::const_reverse_iterator const_reverse_iterator;
 
-  float getASAPStep(const SIRSchedUnit *A) const {
+  unsigned getASAPStep(const SIRSchedUnit *A) const {
     TFMapTy::const_iterator at = SUnitToTF.find(A);
     assert(at != SUnitToTF.end() && "TimeFrame for SU not exist!");
     return at->second.ASAP;
   }
 
-  float getALAPStep(const SIRSchedUnit *A) const {
+  unsigned getALAPStep(const SIRSchedUnit *A) const {
     TFMapTy::const_iterator at = SUnitToTF.find(A);
     assert(at != SUnitToTF.end() && "TimeFrame for SU not exist!");
     return at->second.ALAP;
@@ -122,7 +122,7 @@ public:
   void lengthenCriticalPath() { CriticalPathEnd += 1; }
   void shortenCriticalPath() { CriticalPathEnd -= 1; }
 
-  float getCriticalPathLength() {
+  unsigned getCriticalPathLength() {
     assert(CriticalPathEnd > EntrySlot && "CriticalPathLength not available!");
     return CriticalPathEnd - EntrySlot;
   }
