+++ lib/SIR/SIRSchedGraph.cpp
@@ -25,29 +25,29 @@
 using namespace llvm;
 
 SIRSchedUnit::SIRSchedUnit(unsigned Idx, Type T, BasicBlock *BB)
-  : II(0), Schedule(0.0), Idx(Idx), IsScheduled(false),
+  : II(0), Schedule(0), Idx(Idx), IsScheduled(false),
   T(T), BB(BB), SeqOp(0), CombOp(0) {
   assert(T == SIRSchedUnit::Entry || T == SIRSchedUnit::Exit ||
          T == SIRSchedUnit::BlockEntry && "Unexpected Type for Virtual SUnit!");
 
-  this->Latency = 0.0;
+  this->Latency = 0;
 }
 
 SIRSchedUnit::SIRSchedUnit(unsigned Idx, Type T, BasicBlock *BB, SIRSeqOp *SeqOp)
-  : II(0), Schedule(0.0), Idx(Idx), IsScheduled(false), T(T), BB(BB), SeqOp(SeqOp),
+  : II(0), Schedule(0), Idx(Idx), IsScheduled(false), T(T), BB(BB), SeqOp(SeqOp),
   CombOp(0) {
   assert(T == SIRSchedUnit::SlotTransition || T == SIRSchedUnit::SeqSU ||
          T == SIRSchedUnit::PHI && "Unexpected Type for SeqOp SUnit!");
 
-  this->Latency = 1.0;
+  this->Latency = 1;
 }
 
 SIRSchedUnit::SIRSchedUnit(unsigned Idx, Type T, BasicBlock *BB, Instruction *CombOp)
-  : II(0), Schedule(0.0), Idx(Idx), IsScheduled(false), T(T), BB(BB), SeqOp(0),
+  : II(0), Schedule(0), Idx(Idx), IsScheduled(false), T(T), BB(BB), SeqOp(0),
   CombOp(CombOp) {
   assert(T == SIRSchedUnit::CombSU && "Unexpected Type for CombOp SUnit!");
 
-  this->Latency = 0.0;
+  this->Latency = 0;
 }
 
 SIRSchedUnit::SIRSchedUnit() : Idx(0), T(Invalid), BB(0), SeqOp(0) {}
@@ -107,13 +107,13 @@ SIRDep SIRSchedUnit::EdgeBundle::getEdge(unsigned II) const {
   assert(!Edges.empty() && "Unexpected empty edge bundle!");
 
   SIRDep CurEdge = Edges.front();
-  float CurLatency = CurEdge.getLatency(II);
+  unsigned CurLatency = CurEdge.getLatency(II);
 
   for (unsigned I = 1, E = Edges.size(); I != E; ++I) {
     SIRDep NewEdge = Edges[I];
 
     // Find the edge of II with biggest latency.
-    float NewLatency = NewEdge.getLatency(II);
+    unsigned NewLatency = NewEdge.getLatency(II);
     if (NewLatency > CurLatency) {
       CurLatency = NewLatency;
       CurEdge = NewEdge;
